Crate google_drive3

This documentation was generated from drive crate version 1.0.0+20160901, where 20160901 is the exact revision of the drive:v3 schema built by the mako code generator v1.0.0.

Everything else about the drive v3 API can be found at the official documentation site. The original source code is on github.

Features

Handle the following Resources with ease from the central hub ...

about

get

changes

get start page token, list and watch

channels

stop

comments

create, delete, get, list and update

files

copy, create, delete, empty trash, export, generate ids, get, list, update and watch

permissions

create, delete, get, list and update

replies

create, delete, get, list and update

revisions

delete, get, list and update

Upload supported by ...

update files

create files

Download supported by ...

watch files

export files

get revisions

get files

Subscription supported by ...

watch changes

watch files

list changes

create files

get files

Not what you are looking for ? Find all other Google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

creates Method Builders which in turn allow access to individual Call Builders

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

All structures are marked with applicable traits to further categorize them and ease browsing.

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.files().watch(...).doit()
let r = hub.files().empty_trash(...).doit()
let r = hub.files().generate_ids(...).doit()
let r = hub.files().copy(...).doit()
let r = hub.files().create(...).doit()
let r = hub.files().list(...).doit()
let r = hub.files().delete(...).doit()
let r = hub.files().update(...).doit()
let r = hub.files().get(...).doit()
let r = hub.files().export(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
google-drive3 = "*"

A complete example

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_drive3 as drive3;
use drive3::File;
use drive3::{Result, Error};
use std::fs;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use drive3::Drive;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Drive::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = File::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `upload_resumable(...)`.
// Values shown here are possibly random and not representative !
let result = hub.files().update(req, "fileId")
             .use_content_as_indexable_text(false)
             .remove_parents("eirmod")
             .ocr_language("consetetur")
             .keep_revision_forever(false)
             .add_parents("sed")
             .upload_resumable(fs::File::open("file.ext").unwrap(), "application/octet-stream".parse().unwrap());
 
match result {
    Err(e) => match e {
        // The Error enum provides details about what exactly happened.
        // You can also just use its `Debug`, `Display` or `Error` traits
         Error::HttpError(_)
        |Error::MissingAPIKey
        |Error::MissingToken(_)
        |Error::Cancelled
        |Error::UploadSizeLimitExceeded(_, _)
        |Error::Failure(_)
        |Error::BadRequest(_)
        |Error::FieldClash(_)
        |Error::JsonDecodeError(_, _) => println!("{}", e),
    },
    Ok(res) => println!("Success: {:?}", res),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

Uploads and Downloads

If a method supports downloads, the response body, which is part of the Result, should be read by you to obtain the media. If such a method also supports a Response Result, it will return that by default. You can see it as meta-data for the actual media. To trigger a media download, you will have to set up the builder by making this call: .param("alt", "media").

Methods supporting uploads can do so using up to 2 different protocols: simple and resumable. The distinctiveness of each is represented by customized doit(...) methods, which are then named upload(...) and upload_resumable(...) respectively.

Customization and Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

Optional Parts in Server-Requests

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifiable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are moved

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Structs

About
Information about the user, the user's Drive, and system capabilities.

AboutGetCall
Gets information about the user, the user's Drive, and system capabilities.

AboutMethods
A builder providing access to all methods supported on about resources. It is not used directly, but through the Drive hub.

AboutStorageQuota
The user's storage quota limits and usage. All fields are measured in bytes.

Change
A change to a file.

ChangeGetStartPageTokenCall
Gets the starting pageToken for listing future changes.

ChangeList
A list of changes for a user.

ChangeListCall
Lists changes for a user.

ChangeMethods
A builder providing access to all methods supported on change resources. It is not used directly, but through the Drive hub.

ChangeWatchCall
Subscribes to changes for a user.

Channel
An notification channel used to watch for resource changes.

ChannelMethods
A builder providing access to all methods supported on channel resources. It is not used directly, but through the Drive hub.

ChannelStopCall
Stop watching resources through this channel

Comment
A comment on a file.

CommentCreateCall
Creates a new comment on a file.

CommentDeleteCall
Deletes a comment.

CommentGetCall
Gets a comment by ID.

CommentList
A list of comments on a file.

CommentListCall
Lists a file's comments.

CommentMethods
A builder providing access to all methods supported on comment resources. It is not used directly, but through the Drive hub.

CommentQuotedFileContent
The file content to which the comment refers, typically within the anchor region. For a text file, for example, this would be the text at the location of the comment.

CommentUpdateCall
Updates a comment with patch semantics.

DefaultDelegate
A delegate with a conservative default implementation, which is used if no other delegate is set.

Drive
Central instance to access all Drive related resource activities

ErrorResponse
A utility to represent detailed errors we might see in case there are BadRequests. The latter happen if the sent parameters or request structures are unsound

File
The metadata for a file.

FileCapabilities
Capabilities the current user has on the file.

FileContentHints
Additional information about the content of the file. These fields are never populated in responses.

FileContentHintsThumbnail
A thumbnail for the file. This will only be used if Drive cannot generate a standard thumbnail.

FileCopyCall
Creates a copy of a file and applies any requested updates with patch semantics.

FileCreateCall
Creates a new file.

FileDeleteCall
Permanently deletes a file owned by the user without moving it to the trash. If the target is a folder, all descendants owned by the user are also deleted.

FileEmptyTrashCall
Permanently deletes all of the user's trashed files.

FileExportCall
Exports a Google Doc to the requested MIME type and returns the exported content.

FileGenerateIdCall
Generates a set of file IDs which can be provided in create requests.

FileGetCall
Gets a file's metadata or content by ID.

FileImageMediaMetadata
Additional metadata about image media, if available.

FileImageMediaMetadataLocation
Geographic location information stored in the image.

FileList
A list of files.

FileListCall
Lists or searches files.

FileMethods
A builder providing access to all methods supported on file resources. It is not used directly, but through the Drive hub.

FileUpdateCall
Updates a file's metadata and/or content with patch semantics.

FileVideoMediaMetadata
Additional metadata about video media. This may not be available immediately upon upload.

FileWatchCall
Subscribes to changes to a file

GeneratedIds
A list of generated file IDs which can be provided in create requests.

MethodInfo
Contains information about an API request.

MultiPartReader
Provides a Read interface that converts multiple parts into the protocol identified by RFC2387. Note: This implementation is just as rich as it needs to be to perform uploads to google APIs, and might not be a fully-featured implementation.

Permission
A permission for a file. A permission grants a user, group, domain or the world access to a file or a folder hierarchy.

PermissionCreateCall
Creates a permission for a file.

PermissionDeleteCall
Deletes a permission.

PermissionGetCall
Gets a permission by ID.

PermissionList
A list of permissions for a file.

PermissionListCall
Lists a file's permissions.

PermissionMethods
A builder providing access to all methods supported on permission resources. It is not used directly, but through the Drive hub.

PermissionUpdateCall
Updates a permission with patch semantics.

Reply
A reply to a comment on a file.

ReplyCreateCall
Creates a new reply to a comment.

ReplyDeleteCall
Deletes a reply.

ReplyGetCall
Gets a reply by ID.

ReplyList
A list of replies to a comment on a file.

ReplyListCall
Lists a comment's replies.

ReplyMethods
A builder providing access to all methods supported on reply resources. It is not used directly, but through the Drive hub.

ReplyUpdateCall
Updates a reply with patch semantics.

Revision
The metadata for a revision to a file.

RevisionDeleteCall
Permanently deletes a revision. This method is only applicable to files with binary content in Drive.

RevisionGetCall
Gets a revision's metadata or content by ID.

RevisionList
A list of revisions of a file.

RevisionListCall
Lists a file's revisions.

RevisionMethods
A builder providing access to all methods supported on revision resources. It is not used directly, but through the Drive hub.

RevisionUpdateCall
Updates a revision with patch semantics.

StartPageToken
There is no detailed description.

User
Information about a Drive user.

Enums

Error

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.

Traits

CallBuilder
Identifies types which represent builders for a particular resource method

Delegate
A trait specifying functionality to help controlling any request performed by the API. The trait has a conservative default implementation.

Hub
Identifies the Hub. There is only one per library, this trait is supposed to make intended use more explicit. The hub allows to access all resource methods more easily.

MethodsBuilder
Identifies types for building methods of a particular resource type

NestedType
Identifies types which are only used by other types internally. They have no special meaning, this trait just marks them for completeness.

Part
Identifies types which are only used as part of other types, which usually are carrying the Resource trait.

ReadSeek
A utility to specify reader types which provide seeking capabilities too

RequestValue
Identifies types which are used in API requests.

Resource
Identifies types which can be inserted and deleted. Types with this trait are most commonly used by clients of this API.

ResponseResult
Identifies types which are used in API responses.

ToParts
A trait for all types that can convert themselves into a parts string

Functions

remove_json_null_values

Type Definitions

Result
A universal result type used as return for all calls.