Struct google_drive3::api::File

pub struct File {
    pub app_properties: Option<HashMap<String, String>>,
    pub capabilities: Option<FileCapabilities>,
    pub content_hints: Option<FileContentHints>,
    pub content_restrictions: Option<Vec<ContentRestriction>>,
    pub copy_requires_writer_permission: Option<bool>,
    pub created_time: Option<String>,
    pub description: Option<String>,
    pub drive_id: Option<String>,
    pub explicitly_trashed: Option<bool>,
    pub export_links: Option<HashMap<String, String>>,
    pub file_extension: Option<String>,
    pub folder_color_rgb: Option<String>,
    pub full_file_extension: Option<String>,
    pub has_augmented_permissions: Option<bool>,
    pub has_thumbnail: Option<bool>,
    pub head_revision_id: Option<String>,
    pub icon_link: Option<String>,
    pub id: Option<String>,
    pub image_media_metadata: Option<FileImageMediaMetadata>,
    pub is_app_authorized: Option<bool>,
    pub kind: Option<String>,
    pub last_modifying_user: Option<User>,
    pub md5_checksum: Option<String>,
    pub mime_type: Option<String>,
    pub modified_by_me: Option<bool>,
    pub modified_by_me_time: Option<String>,
    pub modified_time: Option<String>,
    pub name: Option<String>,
    pub original_filename: Option<String>,
    pub owned_by_me: Option<bool>,
    pub owners: Option<Vec<User>>,
    pub parents: Option<Vec<String>>,
    pub permission_ids: Option<Vec<String>>,
    pub permissions: Option<Vec<Permission>>,
    pub properties: Option<HashMap<String, String>>,
    pub quota_bytes_used: Option<String>,
    pub shared: Option<bool>,
    pub shared_with_me_time: Option<String>,
    pub sharing_user: Option<User>,
    pub shortcut_details: Option<FileShortcutDetails>,
    pub size: Option<String>,
    pub spaces: Option<Vec<String>>,
    pub starred: Option<bool>,
    pub team_drive_id: Option<String>,
    pub thumbnail_link: Option<String>,
    pub thumbnail_version: Option<String>,
    pub trashed: Option<bool>,
    pub trashed_time: Option<String>,
    pub trashing_user: Option<User>,
    pub version: Option<String>,
    pub video_media_metadata: Option<FileVideoMediaMetadata>,
    pub viewed_by_me: Option<bool>,
    pub viewed_by_me_time: Option<String>,
    pub viewers_can_copy_content: Option<bool>,
    pub web_content_link: Option<String>,
    pub web_view_link: Option<String>,
    pub writers_can_share: Option<bool>,
}

The metadata for a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

copy files (request|response)

create files (request|response)

delete files (none)

empty trash files (none)

export files (none)

generate ids files (none)

get files (response)

list files (none)

update files (request|response)

watch files (none)

Fields
app_properties: Option<HashMap<String, String>>

A collection of arbitrary key-value pairs which are private to the requesting app. Entries with null values are cleared in update and copy requests. These properties can only be retrieved using an authenticated request. An authenticated request uses an access token obtained with a OAuth 2 client ID. You cannot use an API key to retrieve private properties.

capabilities: Option<FileCapabilities>

Capabilities the current user has on this file. Each capability corresponds to a fine-grained action that a user may take.

content_hints: Option<FileContentHints>

Additional information about the content of the file. These fields are never populated in responses.

content_restrictions: Option<Vec<ContentRestriction>>

Restrictions for accessing the content of the file. Only populated if such a restriction exists.

copy_requires_writer_permission: Option<bool>

Whether the options to copy, print, or download this file, should be disabled for readers and commenters.

created_time: Option<String>

The time at which the file was created (RFC 3339 date-time).

description: Option<String>

A short description of the file.

drive_id: Option<String>

ID of the shared drive the file resides in. Only populated for items in shared drives.

explicitly_trashed: Option<bool>

Whether the file has been explicitly trashed, as opposed to recursively trashed from a parent folder.

export_links: Option<HashMap<String, String>>

Links for exporting Docs Editors files to specific formats.

file_extension: Option<String>

The final component of fullFileExtension. This is only available for files with binary content in Google Drive.

folder_color_rgb: Option<String>

The color for a folder as an RGB hex string. The supported colors are published in the folderColorPalette field of the About resource. If an unsupported color is specified, the closest color in the palette will be used instead.

full_file_extension: Option<String>

The full file extension extracted from the name field. May contain multiple concatenated extensions, such as "tar.gz". This is only available for files with binary content in Google Drive. This is automatically updated when the name field changes, however it is not cleared if the new name does not contain a valid extension.

has_augmented_permissions: Option<bool>

Whether there are permissions directly on this file. This field is only populated for items in shared drives.

has_thumbnail: Option<bool>

Whether this file has a thumbnail. This does not indicate whether the requesting app has access to the thumbnail. To check access, look for the presence of the thumbnailLink field.

head_revision_id: Option<String>

The ID of the file's head revision. This is currently only available for files with binary content in Google Drive.

icon_link: Option<String>

A static, unauthenticated link to the file's icon.

id: Option<String>

The ID of the file.

image_media_metadata: Option<FileImageMediaMetadata>

Additional metadata about image media, if available.

is_app_authorized: Option<bool>

Whether the file was created or opened by the requesting app.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "drive#file".

last_modifying_user: Option<User>

The last user to modify the file.

md5_checksum: Option<String>

The MD5 checksum for the content of the file. This is only applicable to files with binary content in Google Drive.

mime_type: Option<String>

The MIME type of the file. Google Drive will attempt to automatically detect an appropriate value from uploaded content if no value is provided. The value cannot be changed unless a new revision is uploaded. If a file is created with a Google Doc MIME type, the uploaded content will be imported if possible. The supported import formats are published in the About resource.

modified_by_me: Option<bool>

Whether the file has been modified by this user.

modified_by_me_time: Option<String>

The last time the file was modified by the user (RFC 3339 date-time).

modified_time: Option<String>

The last time the file was modified by anyone (RFC 3339 date-time). Note that setting modifiedTime will also update modifiedByMeTime for the user.

name: Option<String>

The name of the file. This is not necessarily unique within a folder. Note that for immutable items such as the top level folders of shared drives, My Drive root folder, and Application Data folder the name is constant.

original_filename: Option<String>

The original filename of the uploaded content if available, or else the original value of the name field. This is only available for files with binary content in Google Drive.

owned_by_me: Option<bool>

Whether the user owns the file. Not populated for items in shared drives.

owners: Option<Vec<User>>

The owners of the file. Currently, only certain legacy files may have more than one owner. Not populated for items in shared drives.

parents: Option<Vec<String>>

The IDs of the parent folders which contain the file. If not specified as part of a create request, the file will be placed directly in the user's My Drive folder. If not specified as part of a copy request, the file will inherit any discoverable parents of the source file. Update requests must use the addParents and removeParents parameters to modify the parents list.

permission_ids: Option<Vec<String>>

List of permission IDs for users with access to this file.

permissions: Option<Vec<Permission>>

The full list of permissions for the file. This is only available if the requesting user can share the file. Not populated for items in shared drives.

properties: Option<HashMap<String, String>>

A collection of arbitrary key-value pairs which are visible to all apps. Entries with null values are cleared in update and copy requests.

quota_bytes_used: Option<String>

The number of storage quota bytes used by the file. This includes the head revision as well as previous revisions with keepForever enabled.

shared: Option<bool>

Whether the file has been shared. Not populated for items in shared drives.

shared_with_me_time: Option<String>

The time at which the file was shared with the user, if applicable (RFC 3339 date-time).

sharing_user: Option<User>

The user who shared the file with the requesting user, if applicable.

shortcut_details: Option<FileShortcutDetails>

Shortcut file details. Only populated for shortcut files, which have the mimeType field set to application/vnd.google-apps.shortcut.

size: Option<String>

The size of the file's content in bytes. This is applicable to binary files in Google Drive and Google Docs files.

spaces: Option<Vec<String>>

The list of spaces which contain the file. The currently supported values are 'drive', 'appDataFolder' and 'photos'.

starred: Option<bool>

Whether the user has starred the file.

team_drive_id: Option<String>

Deprecated - use driveId instead.

thumbnail_link: Option<String>

A short-lived link to the file's thumbnail, if available. Typically lasts on the order of hours. Only populated when the requesting app can access the file's content. If the file isn't shared publicly, the URL returned in Files.thumbnailLink must be fetched using a credentialed request.

thumbnail_version: Option<String>

The thumbnail version for use in thumbnail cache invalidation.

trashed: Option<bool>

Whether the file has been trashed, either explicitly or from a trashed parent folder. Only the owner may trash a file. The trashed item is excluded from all files.list responses returned for any user who does not own the file. However, all users with access to the file can see the trashed item metadata in an API response. All users with access can copy, download, export, and share the file.

trashed_time: Option<String>

The time that the item was trashed (RFC 3339 date-time). Only populated for items in shared drives.

trashing_user: Option<User>

If the file has been explicitly trashed, the user who trashed it. Only populated for items in shared drives.

version: Option<String>

A monotonically increasing version number for the file. This reflects every change made to the file on the server, even those not visible to the user.

video_media_metadata: Option<FileVideoMediaMetadata>

Additional metadata about video media. This may not be available immediately upon upload.

viewed_by_me: Option<bool>

Whether the file has been viewed by this user.

viewed_by_me_time: Option<String>

The last time the file was viewed by the user (RFC 3339 date-time).

viewers_can_copy_content: Option<bool>

Deprecated - use copyRequiresWriterPermission instead.

web_content_link: Option<String>

A link for downloading the content of the file in a browser. This is only available for files with binary content in Google Drive.

web_view_link: Option<String>

A link for opening the file in a relevant Google editor or viewer in a browser.

writers_can_share: Option<bool>

Whether users with only writer permission can modify the file's permissions. Not populated for items in shared drives.