Struct google_drive3::api::FileImageMediaMetadata

pub struct FileImageMediaMetadata {
    pub aperture: Option<f32>,
    pub camera_make: Option<String>,
    pub camera_model: Option<String>,
    pub color_space: Option<String>,
    pub exposure_bias: Option<f32>,
    pub exposure_mode: Option<String>,
    pub exposure_time: Option<f32>,
    pub flash_used: Option<bool>,
    pub focal_length: Option<f32>,
    pub height: Option<i32>,
    pub iso_speed: Option<i32>,
    pub lens: Option<String>,
    pub location: Option<FileImageMediaMetadataLocation>,
    pub max_aperture_value: Option<f32>,
    pub metering_mode: Option<String>,
    pub rotation: Option<i32>,
    pub sensor: Option<String>,
    pub subject_distance: Option<i32>,
    pub time: Option<String>,
    pub white_balance: Option<String>,
    pub width: Option<i32>,
}

Additional metadata about image media, if available.

This type is not used in any activity, and only used as part of another schema.

Fields
aperture: Option<f32>

The aperture used to create the photo (f-number).

camera_make: Option<String>

The make of the camera used to create the photo.

camera_model: Option<String>

The model of the camera used to create the photo.

color_space: Option<String>

The color space of the photo.

exposure_bias: Option<f32>

The exposure bias of the photo (APEX value).

exposure_mode: Option<String>

The exposure mode used to create the photo.

exposure_time: Option<f32>

The length of the exposure, in seconds.

flash_used: Option<bool>

Whether a flash was used to create the photo.

focal_length: Option<f32>

The focal length used to create the photo, in millimeters.

height: Option<i32>

The height of the image in pixels.

iso_speed: Option<i32>

The ISO speed used to create the photo.

lens: Option<String>

The lens used to create the photo.

location: Option<FileImageMediaMetadataLocation>

Geographic location information stored in the image.

max_aperture_value: Option<f32>

The smallest f-number of the lens at the focal length used to create the photo (APEX value).

metering_mode: Option<String>

The metering mode used to create the photo.

rotation: Option<i32>

The number of clockwise 90 degree rotations applied from the image's original orientation.

sensor: Option<String>

The type of sensor used to create the photo.

subject_distance: Option<i32>

The distance to the subject of the photo, in meters.

time: Option<String>

The date and time the photo was taken (EXIF DateTime).

white_balance: Option<String>

The white balance mode used to create the photo.

width: Option<i32>

The width of the image in pixels.