Struct google_drive3::Comment

pub struct Comment {
    pub resolved: Option<bool>,
    pub kind: Option<String>,
    pub modified_time: Option<String>,
    pub author: Option<User>,
    pub deleted: Option<bool>,
    pub quoted_file_content: Option<CommentQuotedFileContent>,
    pub html_content: Option<String>,
    pub anchor: Option<String>,
    pub content: Option<String>,
    pub replies: Option<Vec<Reply>>,
    pub created_time: Option<String>,
    pub id: Option<String>,
}

A comment on a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update comments (request|response)

create comments (request|response)

get comments (response)

list comments (none)

delete comments (none)

Fields

resolved

Whether the comment has been resolved by one of its replies.

kind

This is always drive#comment.

modified_time

The last time the comment or any of its replies was modified (RFC 3339 date-time).

author

The user who created the comment.

deleted

Whether the comment has been deleted. A deleted comment has no content.

quoted_file_content

The file content to which the comment refers, typically within the anchor region. For a text file, for example, this would be the text at the location of the comment.

html_content

The content of the comment with HTML formatting.

anchor

A region of the document represented as a JSON string. See anchor documentation for details on how to define and interpret anchor properties.

content

The plain text content of the comment. This field is used for setting the content, while htmlContent should be displayed.

replies

The full list of replies to the comment in chronological order.

created_time

The time at which the comment was created (RFC 3339 date-time).

id

The ID of the comment.