Struct google_drive3::File

pub struct File {
    pub mime_type: Option<String>,
    pub modified_by_me_time: Option<String>,
    pub thumbnail_link: Option<String>,
    pub explicitly_trashed: Option<bool>,
    pub is_app_authorized: Option<bool>,
    pub writers_can_share: Option<bool>,
    pub owned_by_me: Option<bool>,
    pub viewed_by_me_time: Option<String>,
    pub id: Option<String>,
    pub sharing_user: Option<User>,
    pub size: Option<String>,
    pub video_media_metadata: Option<FileVideoMediaMetadata>,
    pub last_modifying_user: Option<User>,
    pub folder_color_rgb: Option<String>,
    pub app_properties: Option<HashMap<String, String>>,
    pub capabilities: Option<FileCapabilities>,
    pub web_view_link: Option<String>,
    pub version: Option<String>,
    pub parents: Option<Vec<String>>,
    pub md5_checksum: Option<String>,
    pub shared: Option<bool>,
    pub icon_link: Option<String>,
    pub full_file_extension: Option<String>,
    pub original_filename: Option<String>,
    pub image_media_metadata: Option<FileImageMediaMetadata>,
    pub description: Option<String>,
    pub modified_time: Option<String>,
    pub viewed_by_me: Option<bool>,
    pub kind: Option<String>,
    pub created_time: Option<String>,
    pub quota_bytes_used: Option<String>,
    pub properties: Option<HashMap<String, String>>,
    pub shared_with_me_time: Option<String>,
    pub permissions: Option<Vec<Permission>>,
    pub viewers_can_copy_content: Option<bool>,
    pub owners: Option<Vec<User>>,
    pub name: Option<String>,
    pub web_content_link: Option<String>,
    pub spaces: Option<Vec<String>>,
    pub trashed: Option<bool>,
    pub content_hints: Option<FileContentHints>,
    pub file_extension: Option<String>,
    pub starred: Option<bool>,
    pub head_revision_id: Option<String>,
}

The metadata for a file.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

watch files (none)

empty trash files (none)

generate ids files (none)

copy files (request|response)

create files (request|response)

list files (none)

delete files (none)

update files (request|response)

get files (response)

export files (none)

Fields

mime_type

The MIME type of the file. Drive will attempt to automatically detect an appropriate value from uploaded content if no value is provided. The value cannot be changed unless a new revision is uploaded. If a file is created with a Google Doc MIME type, the uploaded content will be imported if possible. The supported import formats are published in the About resource.

modified_by_me_time

The last time the file was modified by the user (RFC 3339 date-time).

thumbnail_link

A short-lived link to the file's thumbnail, if available. Typically lasts on the order of hours.

explicitly_trashed

Whether the file has been explicitly trashed, as opposed to recursively trashed from a parent folder.

is_app_authorized

Whether the file was created or opened by the requesting app.

writers_can_share

Whether users with only writer permission can modify the file's permissions.

owned_by_me

Whether the user owns the file.

viewed_by_me_time

The last time the file was viewed by the user (RFC 3339 date-time).

id

The ID of the file.

sharing_user

The user who shared the file with the requesting user, if applicable.

size

The size of the file's content in bytes. This is only applicable to files with binary content in Drive.

video_media_metadata

Additional metadata about video media. This may not be available immediately upon upload.

last_modifying_user

The last user to modify the file.

folder_color_rgb

The color for a folder as an RGB hex string. The supported colors are published in the folderColorPalette field of the About resource. If an unsupported color is specified, the closest color in the palette will be used instead.

app_properties

A collection of arbitrary key-value pairs which are private to the requesting app. Entries with null values are cleared in update and copy requests.

capabilities

Capabilities the current user has on the file.

web_view_link

A link for opening the file in a relevant Google editor or viewer in a browser.

version

A monotonically increasing version number for the file. This reflects every change made to the file on the server, even those not visible to the user.

parents

The IDs of the parent folders which contain the file. If not specified as part of a create request, the file will be placed directly in the My Drive folder. Update requests must use the addParents and removeParents parameters to modify the values.

md5_checksum

The MD5 checksum for the content of the file. This is only applicable to files with binary content in Drive.

shared

Whether the file has been shared.

icon_link

A static, unauthenticated link to the file's icon.

full_file_extension

The full file extension extracted from the name field. May contain multiple concatenated extensions, such as "tar.gz". This is only available for files with binary content in Drive. This is automatically updated when the name field changes, however it is not cleared if the new name does not contain a valid extension.

original_filename

The original filename of the uploaded content if available, or else the original value of the name field. This is only available for files with binary content in Drive.

image_media_metadata

Additional metadata about image media, if available.

description

A short description of the file.

modified_time

The last time the file was modified by anyone (RFC 3339 date-time). Note that setting modifiedTime will also update modifiedByMeTime for the user.

viewed_by_me

Whether the file has been viewed by this user.

kind

This is always drive#file.

created_time

The time at which the file was created (RFC 3339 date-time).

quota_bytes_used

The number of storage quota bytes used by the file. This includes the head revision as well as previous revisions with keepForever enabled.

properties

A collection of arbitrary key-value pairs which are visible to all apps. Entries with null values are cleared in update and copy requests.

shared_with_me_time

The time at which the file was shared with the user, if applicable (RFC 3339 date-time).

permissions

The full list of permissions for the file. This is only available if the requesting user can share the file.

viewers_can_copy_content

Whether users with only reader or commenter permission can copy the file's content. This affects copy, download, and print operations.

owners

The owners of the file. Currently, only certain legacy files may have more than one owner.

name

The name of the file. This is not necessarily unique within a folder.

web_content_link

A link for downloading the content of the file in a browser. This is only available for files with binary content in Drive.

spaces

The list of spaces which contain the file. The currently supported values are 'drive', 'appDataFolder' and 'photos'.

trashed

Whether the file has been trashed, either explicitly or from a trashed parent folder. Only the owner may trash a file, and other users cannot see files in the owner's trash.

content_hints

Additional information about the content of the file. These fields are never populated in responses.

file_extension

The final component of fullFileExtension. This is only available for files with binary content in Drive.

starred

Whether the user has starred the file.

head_revision_id

The ID of the file's head revision. This is currently only available for files with binary content in Drive.