Struct google_drive3::FileCopyCall

pub struct FileCopyCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Creates a copy of a file and applies any requested updates with patch semantics.

A builder for the copy method supported by a file resource. It is not used directly, but through a FileMethods instance.

Example

Instantiate a resource method builder

use drive3::File;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = File::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.files().copy(req, "fileId")
             .ocr_language("justo")
             .keep_revision_forever(true)
             .ignore_default_visibility(true)
             .doit();

Methods

impl<'a, C, A> FileCopyCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, File)>

Perform the operation you have build so far.

fn request(self, new_value: File) -> FileCopyCall<'a, C, A>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn file_id(self, new_value: &str) -> FileCopyCall<'a, C, A>

The ID of the file.

Sets the file id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn ocr_language(self, new_value: &str) -> FileCopyCall<'a, C, A>

A language hint for OCR processing during image import (ISO 639-1 code).

Sets the ocr language query property to the given value.

fn keep_revision_forever(self, new_value: bool) -> FileCopyCall<'a, C, A>

Whether to set the 'keepForever' field in the new head revision. This is only applicable to files with binary content in Drive.

Sets the keep revision forever query property to the given value.

fn ignore_default_visibility(self, new_value: bool) -> FileCopyCall<'a, C, A>

Whether to ignore the domain's default visibility settings for the created file. Domain administrators can choose to make all uploaded files visible to the domain by default; this parameter bypasses that behavior for the request. Permissions are still inherited from parent folders.

Sets the ignore default visibility query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> FileCopyCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> FileCopyCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> FileCopyCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Full.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.