Struct google_drive3::FileImageMediaMetadata

pub struct FileImageMediaMetadata {
    pub exposure_bias: Option<f32>,
    pub exposure_time: Option<f32>,
    pub max_aperture_value: Option<f32>,
    pub color_space: Option<String>,
    pub height: Option<i32>,
    pub lens: Option<String>,
    pub aperture: Option<f32>,
    pub rotation: Option<i32>,
    pub white_balance: Option<String>,
    pub camera_model: Option<String>,
    pub flash_used: Option<bool>,
    pub camera_make: Option<String>,
    pub focal_length: Option<f32>,
    pub exposure_mode: Option<String>,
    pub subject_distance: Option<i32>,
    pub width: Option<i32>,
    pub metering_mode: Option<String>,
    pub location: Option<FileImageMediaMetadataLocation>,
    pub time: Option<String>,
    pub iso_speed: Option<i32>,
    pub sensor: Option<String>,
}

Additional metadata about image media, if available.

This type is not used in any activity, and only used as part of another schema.

Fields

exposure_bias

The exposure bias of the photo (APEX value).

exposure_time

The length of the exposure, in seconds.

max_aperture_value

The smallest f-number of the lens at the focal length used to create the photo (APEX value).

color_space

The color space of the photo.

height

The height of the image in pixels.

lens

The lens used to create the photo.

aperture

The aperture used to create the photo (f-number).

rotation

The rotation in clockwise degrees from the image's original orientation.

white_balance

The white balance mode used to create the photo.

camera_model

The model of the camera used to create the photo.

flash_used

Whether a flash was used to create the photo.

camera_make

The make of the camera used to create the photo.

focal_length

The focal length used to create the photo, in millimeters.

exposure_mode

The exposure mode used to create the photo.

subject_distance

The distance to the subject of the photo, in meters.

width

The width of the image in pixels.

metering_mode

The metering mode used to create the photo.

location

Geographic location information stored in the image.

time

The date and time the photo was taken (EXIF DateTime).

iso_speed

The ISO speed used to create the photo.

sensor

The type of sensor used to create the photo.