Struct google_drive3::FileMethods

pub struct FileMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on file resources. It is not used directly, but through the Drive hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_drive3 as drive3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use drive3::Drive;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Drive::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `copy(...)`, `create(...)`, `delete(...)`, `empty_trash(...)`, `export(...)`, `generate_ids(...)`, `get(...)`, `list(...)`, `update(...)` and `watch(...)`
// to build up your call.
let rb = hub.files();

Methods

impl<'a, C, A> FileMethods<'a, C, A>

fn update(&self, request: File, file_id: &str) -> FileUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a file's metadata and/or content with patch semantics.

Arguments

request - No description provided.

fileId - The ID of the file.

fn watch(&self, request: Channel, file_id: &str) -> FileWatchCall<'a, C, A>

Create a builder to help you perform the following task:

Subscribes to changes to a file

Arguments

request - No description provided.

fileId - The ID of the file.

fn copy(&self, request: File, file_id: &str) -> FileCopyCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a copy of a file and applies any requested updates with patch semantics.

Arguments

request - No description provided.

fileId - The ID of the file.

fn export(&self, file_id: &str, mime_type: &str) -> FileExportCall<'a, C, A>

Create a builder to help you perform the following task:

Exports a Google Doc to the requested MIME type and returns the exported content.

Arguments

fileId - The ID of the file.

mimeType - The MIME type of the format requested for this export.

fn empty_trash(&self) -> FileEmptyTrashCall<'a, C, A>

Create a builder to help you perform the following task:

Permanently deletes all of the user's trashed files.

fn list(&self) -> FileListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists or searches files.

fn generate_ids(&self) -> FileGenerateIdCall<'a, C, A>

Create a builder to help you perform the following task:

Generates a set of file IDs which can be provided in create requests.

fn create(&self, request: File) -> FileCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new file.

Arguments

request - No description provided.

fn delete(&self, file_id: &str) -> FileDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Permanently deletes a file owned by the user without moving it to the trash. If the target is a folder, all descendants owned by the user are also deleted.

Arguments

fileId - The ID of the file.

fn get(&self, file_id: &str) -> FileGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a file's metadata or content by ID.

Arguments

fileId - The ID of the file.