Struct google_drive3::PermissionMethods

pub struct PermissionMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on permission resources. It is not used directly, but through the Drive hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_drive3 as drive3;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use drive3::Drive;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Drive::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `create(...)`, `delete(...)`, `get(...)`, `list(...)` and `update(...)`
// to build up your call.
let rb = hub.permissions();

Methods

impl<'a, C, A> PermissionMethods<'a, C, A>

fn delete(&self, file_id: &str, permission_id: &str) -> PermissionDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a permission.

Arguments

fileId - The ID of the file.

permissionId - The ID of the permission.

fn list(&self, file_id: &str) -> PermissionListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists a file's permissions.

Arguments

fileId - The ID of the file.

fn update(&self, request: Permission, file_id: &str, permission_id: &str) -> PermissionUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a permission with patch semantics.

Arguments

request - No description provided.

fileId - The ID of the file.

permissionId - The ID of the permission.

fn get(&self, file_id: &str, permission_id: &str) -> PermissionGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a permission by ID.

Arguments

fileId - The ID of the file.

permissionId - The ID of the permission.

fn create(&self, request: Permission, file_id: &str) -> PermissionCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a permission for a file.

Arguments

request - No description provided.

fileId - The ID of the file.