Module google_drive3::api

Structs

About
Information about the user, the user's Drive, and system capabilities.

AboutDriveThemes
A list of themes that are supported for shared drives.

AboutGetCall
Gets information about the user, the user's Drive, and system capabilities.

AboutMethods
A builder providing access to all methods supported on about resources. It is not used directly, but through the DriveHub hub.

AboutStorageQuota
The user's storage quota limits and usage. All fields are measured in bytes.

AboutTeamDriveThemes
Deprecated - use driveThemes instead.

Change
A change to a file or shared drive.

ChangeGetStartPageTokenCall
Gets the starting pageToken for listing future changes.

ChangeList
A list of changes for a user.

ChangeListCall
Lists the changes for a user or shared drive.

ChangeMethods
A builder providing access to all methods supported on change resources. It is not used directly, but through the DriveHub hub.

ChangeWatchCall
Subscribes to changes for a user.

Channel
An notification channel used to watch for resource changes.

ChannelMethods
A builder providing access to all methods supported on channel resources. It is not used directly, but through the DriveHub hub.

ChannelStopCall
Stop watching resources through this channel

Comment
A comment on a file.

CommentCreateCall
Creates a new comment on a file.

CommentDeleteCall
Deletes a comment.

CommentGetCall
Gets a comment by ID.

CommentList
A list of comments on a file.

CommentListCall
Lists a file's comments.

CommentMethods
A builder providing access to all methods supported on comment resources. It is not used directly, but through the DriveHub hub.

CommentQuotedFileContent
The file content to which the comment refers, typically within the anchor region. For a text file, for example, this would be the text at the location of the comment.

CommentUpdateCall
Updates a comment with patch semantics.

ContentRestriction
A restriction for accessing the content of the file.

Drive
Representation of a shared drive.

DriveBackgroundImageFile
An image file and cropping parameters from which a background image for this shared drive is set. This is a write only field; it can only be set on drive.drives.update requests that don't set themeId. When specified, all fields of the backgroundImageFile must be set.

DriveCapabilities
Capabilities the current user has on this shared drive.

DriveCreateCall
Creates a new shared drive.

DriveDeleteCall
Permanently deletes a shared drive for which the user is an organizer. The shared drive cannot contain any untrashed items.

DriveGetCall
Gets a shared drive's metadata by ID.

DriveHideCall
Hides a shared drive from the default view.

DriveHub
Central instance to access all DriveHub related resource activities

DriveList
A list of shared drives.

DriveListCall
Lists the user's shared drives.

DriveMethods
A builder providing access to all methods supported on drive resources. It is not used directly, but through the DriveHub hub.

DriveRestrictions
A set of restrictions that apply to this shared drive or items inside this shared drive.

DriveUnhideCall
Restores a shared drive to the default view.

DriveUpdateCall
Updates the metadate for a shared drive.

File
The metadata for a file.

FileCapabilities
Capabilities the current user has on this file. Each capability corresponds to a fine-grained action that a user may take.

FileContentHints
Additional information about the content of the file. These fields are never populated in responses.

FileContentHintsThumbnail
A thumbnail for the file. This will only be used if Google Drive cannot generate a standard thumbnail.

FileCopyCall
Creates a copy of a file and applies any requested updates with patch semantics. Folders cannot be copied.

FileCreateCall
Creates a new file.

FileDeleteCall
Permanently deletes a file owned by the user without moving it to the trash. If the file belongs to a shared drive the user must be an organizer on the parent. If the target is a folder, all descendants owned by the user are also deleted.

FileEmptyTrashCall
Permanently deletes all of the user's trashed files.

FileExportCall
Exports a Google Doc to the requested MIME type and returns the exported content. Please note that the exported content is limited to 10MB.

FileGenerateIdCall
Generates a set of file IDs which can be provided in create or copy requests.

FileGetCall
Gets a file's metadata or content by ID.

FileImageMediaMetadata
Additional metadata about image media, if available.

FileImageMediaMetadataLocation
Geographic location information stored in the image.

FileList
A list of files.

FileListCall
Lists or searches files.

FileMethods
A builder providing access to all methods supported on file resources. It is not used directly, but through the DriveHub hub.

FileShortcutDetails
Shortcut file details. Only populated for shortcut files, which have the mimeType field set to application/vnd.google-apps.shortcut.

FileUpdateCall
Updates a file's metadata and/or content. This method supports patch semantics.

FileVideoMediaMetadata
Additional metadata about video media. This may not be available immediately upon upload.

FileWatchCall
Subscribes to changes to a file

GeneratedIds
A list of generated file IDs which can be provided in create requests.

Permission
A permission for a file. A permission grants a user, group, domain or the world access to a file or a folder hierarchy.

PermissionCreateCall
Creates a permission for a file or shared drive.

PermissionDeleteCall
Deletes a permission.

PermissionGetCall
Gets a permission by ID.

PermissionList
A list of permissions for a file.

PermissionListCall
Lists a file's or shared drive's permissions.

PermissionMethods
A builder providing access to all methods supported on permission resources. It is not used directly, but through the DriveHub hub.

PermissionPermissionDetails
Details of whether the permissions on this shared drive item are inherited or directly on this item. This is an output-only field which is present only for shared drive items.

PermissionTeamDrivePermissionDetails
Deprecated - use permissionDetails instead.

PermissionUpdateCall
Updates a permission with patch semantics.

Reply
A reply to a comment on a file.

ReplyCreateCall
Creates a new reply to a comment.

ReplyDeleteCall
Deletes a reply.

ReplyGetCall
Gets a reply by ID.

ReplyList
A list of replies to a comment on a file.

ReplyListCall
Lists a comment's replies.

ReplyMethods
A builder providing access to all methods supported on reply resources. It is not used directly, but through the DriveHub hub.

ReplyUpdateCall
Updates a reply with patch semantics.

Revision
The metadata for a revision to a file.

RevisionDeleteCall
Permanently deletes a file version. You can only delete revisions for files with binary content in Google Drive, like images or videos. Revisions for other files, like Google Docs or Sheets, and the last remaining file version can't be deleted.

RevisionGetCall
Gets a revision's metadata or content by ID.

RevisionList
A list of revisions of a file.

RevisionListCall
Lists a file's revisions.

RevisionMethods
A builder providing access to all methods supported on revision resources. It is not used directly, but through the DriveHub hub.

RevisionUpdateCall
Updates a revision with patch semantics.

StartPageToken
There is no detailed description.

TeamDrive
Deprecated: use the drive collection instead.

TeamDriveBackgroundImageFile
An image file and cropping parameters from which a background image for this Team Drive is set. This is a write only field; it can only be set on drive.teamdrives.update requests that don't set themeId. When specified, all fields of the backgroundImageFile must be set.

TeamDriveCapabilities
Capabilities the current user has on this Team Drive.

TeamDriveList
A list of Team Drives.

TeamDriveRestrictions
A set of restrictions that apply to this Team Drive or items inside this Team Drive.

TeamdriveCreateCall
Deprecated use drives.create instead.

TeamdriveDeleteCall
Deprecated use drives.delete instead.

TeamdriveGetCall
Deprecated use drives.get instead.

TeamdriveListCall
Deprecated use drives.list instead.

TeamdriveMethods
A builder providing access to all methods supported on teamdrive resources. It is not used directly, but through the DriveHub hub.

TeamdriveUpdateCall
Deprecated use drives.update instead

User
Information about a Drive user.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.